feat : 사용자 그래프로 사용할 데이터 변환 로직 완성

+++ src/main/java/com/urlive/web/dto/log/TrafficByDateRange.java
@@ -2,5 +2,5 @@ package com.urlive.web.dto.log;
 
 import java.time.LocalDate;
 
-public record TrafficByDateRange(LocalDate date, int count) {
+public record TrafficByDateRange(LocalDate date, long count) {
 }
