+++ src/main/java/com/urlive/web/dto/log/TrafficByDevice.java
@@ -4,5 +4,5 @@ import java.time.LocalDate;
 
 public record TrafficByDevice(LocalDate date,
                               String device,
-                              int count) {
+                              long count) {
 }
